Create README.md
The purpose of this is to use as a guide through my journey with understanding the things that I have learned about PowerShell
while also using the commands and implementing them for projects and other uses in the workplace. This cheat sheet will help me
remember all of the commands that I have ever used throughout my PowerShell 1 course in Durham College and will be very useful
for my IT career.